refactor:  Configurations Context
Modified the URL and Configurations to utilize a context rather then a
resources object for forming request.  This will allow for clearner code
since whenever a resource is needed a Context is required to acquire it.
So by eliminating the Resources contruct we can save on a lot of calls by
simply passing the Context around.